Add support for videos in posts
To do this right we had to implement a custom controller for the uploaded
assets. The reason we needed to do this is because apple devices expect
video asset requests to support byte-range requests, and Plug.Static
apparently doesn't handle those correctly.

The custom handler for the assets here is based on
https://medium.com/@miguel.coba/streaming-video-with-phoenix-31cd6c3048e
but tweaked to be a tiny bit more standards compliant. The standard is
is available at
https://datatracker.ietf.org/doc/html/rfc7233#section-4

On the bright side, this means we have much more fine-grained control
over how we serve the assets, and thus how we store them \o/

One thing of note is how iOS devices support the video/quicktime
mime-type out of the box, but other browsers have to use video/mp4 to
play the same video. video/quicktime videos can be played when the
"type" annotation in the <video> tag is set to video/mp4, so we simply
add an additional source with video/mp4 whenever we're dealing with a
video/quicktime video.